refactor(sputnik): invert javy text_encoding package
