+++ scripts/sputnik.polyfills.mjs
@@ -48,8 +48,8 @@ const transformTextEncodingPolyfill = (content) => {
 
 	return `use std::str;
 
-use crate::js::apis::node::text_encoding::javy::impls::Args;
-use crate::js::apis::node::text_encoding::javy::utils::{to_js_error, to_string_lossy};
+use crate::js::apis::node::javy::text_encoding::impls::Args;
+use crate::js::apis::node::javy::text_encoding::utils::{to_js_error, to_string_lossy};
 use crate::{hold, hold_and_release};
 use anyhow::{anyhow, bail, Error, Result};
 use rquickjs::{
@@ -79,11 +79,11 @@ const savePolyfill = async ({ dest, content, transform }) => {
 const resources = [
 	{
 		src: '/bytecodealliance/javy/refs/heads/main/crates/javy/src/apis/text_encoding/text-encoding.js',
-		dest: join(process.cwd(), 'src/sputnik/src/js/apis/node/text_encoding/javy/text-encoding.js')
+		dest: join(process.cwd(), 'src/sputnik/src/js/apis/node/javy/text_encoding/text-encoding.js')
 	},
 	{
 		src: '/bytecodealliance/javy/refs/heads/main/crates/javy/src/apis/text_encoding/mod.rs',
-		dest: join(process.cwd(), 'src/sputnik/src/js/apis/node/text_encoding/javy/polyfill.rs'),
+		dest: join(process.cwd(), 'src/sputnik/src/js/apis/node/javy/text_encoding/polyfill.rs'),
 		transform: transformTextEncodingPolyfill
 	},
 	{
