+++ src/sputnik/src/js/apis/node/mod.rs
@@ -1,8 +1,8 @@
 mod blob;
-mod text_encoding;
+mod javy;
 
 use crate::js::apis::node::blob::init_blob;
-use crate::js::apis::node::text_encoding::init_text_encoding;
+use crate::js::apis::node::javy::init_text_encoding;
 use rquickjs::{Ctx, Error as JsError};
 
 pub fn init_node_apis(ctx: &Ctx) -> Result<(), JsError> {
